Exception handling
hello guys , this is our exception handling code .Exception handling is crucial for the proper functioning of your applications. With exception handling, developers define the steps for addressing compilation or runtime errors to ensure that there are no interruptions of program flow. As a result, applications are more stable and have a better user experience.