Try_catch_finally
This is important concept in core java. In Exception handling this code indicates use of finally block in java .But finally is useful for more than just exception handling — it allows the programmer to avoid having cleanup code accidentally bypassed by a return , continue , or break . Putting cleanup code in a finally block is always a good practice, even when no exceptions are anticipated.