Throws keyword in java
hello guys , lets talk about the throws keyword in java. this is the important keywords that prevent the code exception .this helps developer to understands the use of try catch were excite we needed try catch In future in now process. Exception Handling is one of the effective means to handle runtime errors so that the regular flow of the application can be preserved. Java Exception Handling is a mechanism to handle runtime errors such as ClassNotFoundException, IOException, SQLException, RemoteException, etc.